chnage version to 1.0
